Multiplayer RoffSystem Fixes (#1165)
* [MP] Fix ROFF system using incorrect time value.

The ROFF system previously used engine time instead of module time when setting the time values for trajectories of shared entities. This caused issues when the server module time has been reset on map change, because the game module would misinterpret the trajectories.

* [MP] Fix roff not working on 64 bit builds.

Change type of mVersion in TROFFHeader and TROFF2Header from long to int.

+++ codemp/qcommon/RoffSystem.h
@@ -64,7 +64,7 @@ private:
 	//-------------------------------
 	{
 		char	mHeader[4];				// should match roff_string defined above
-		long	mVersion;				// version num, supported version defined above
+		int		mVersion;				// version num, supported version defined above
 		float	mCount;					// I think this is a float because of a limitation of the roff exporter
 
 	} TROFFHeader;
@@ -81,7 +81,7 @@ private:
 	//-------------------------------
 	{
 		char	mHeader[4];				// should match roff_string defined above
-		long	mVersion;				// version num, supported version defined above
+		int		mVersion;				// version num, supported version defined above
 		int		mCount;					// I think this is a float because of a limitation of the roff exporter
 		int		mFrameRate;				// Frame rate the roff should be played at
 		int		mNumNotes;				// number of notes (null terminated strings) after the roff data
